cast safety

+++ src/application/UserCapabilities.cpp
@@ -39,7 +39,8 @@ void UserCapabilities::findHardwareDetails() {
   }
 
   // Read from pipe
-  while (fgets(buffer.data(), buffer.size(), pipe.get()) != nullptr) {
+  while (fgets(buffer.data(), static_cast<int>(buffer.size()), pipe.get()) !=
+         nullptr) {
     result += buffer.data();
   }
